+++ tidy3d/compat.py
@@ -1,0 +1,8 @@
+"""Compatibility layer for handling differences between package versions."""
+
+try:
+    from xarray.structure import alignment
+except ImportError:
+    from xarray.core import alignment
+
+__all__ = ["alignment"]